+++ CHANGELOG.md
@@ -7,6 +7,8 @@ backwards compatibility.
 ## Unreleased
 
 * Introduce documentation of the build process.
+* Remove amplitude balancing on small chords. This causes weak signal output but
+  guarantees stable loudness on continuous tones.
 
 ## 1.2.0